Reorganise identity-did/verifiable module
