+++ identity-did/src/verifiable/mod.rs
@@ -1,14 +1,15 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-pub use self::document::DocumentSigner;
-pub use self::document::DocumentVerifier;
+pub use self::document_signer::DocumentSigner;
+pub use self::document_verifier::DocumentVerifier;
+pub use self::revocation::Revocation;
 pub use self::properties::VerifiableProperties;
-pub use self::traits::Revocation;
 
-mod document;
+mod document_signer;
+mod document_verifier;
+mod revocation;
 mod properties;
-mod traits;
 
 #[cfg(test)]
 mod tests;
